Added check-module Makefile target.
This checks modules binary artifacts for the presence of RPATH entries.
On Alpine and Debian-based distros we rely on our custom code.  It can not be
used in RPM since $RPM_BUILD_ROOT is cleaned up by default when rpm packages
are built, so instead we rely on built-in rpm rpath checks.